+++ packages/terra-collapsible-menu-view/tests/wdio/collapsible-menu-view-spec.js
@@ -101,6 +101,15 @@ Terra.describeViewports('Unselectable Toggle', ['tiny', 'small', 'medium', 'huge
   });
 });
 
+Terra.describeViewports('Reversed and start aligned', ['tiny', 'small'], () => {
+  it('should display hyperlinks and chevron right icons', () => {
+    browser.url('/raw/tests/cerner-terra-framework-docs/collapsible-menu-view/reverse-start-aligned');
+    Terra.validates.element('reverse start aligned');
+    $('[data-collapsible-menu-toggle]').click();
+    Terra.validates.element('reverse start aligned menu open', { selector: '#root', rules: ignoredA11y });
+  });
+});
+
 // Only test viewports that have collapsed menu items
 Terra.describeViewports('Start Aligned Hidden Open', ['tiny', 'small', 'medium', 'large'], () => {
   it('should display start aligned with menu open', () => {