+++ packages/terra-application-layout/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 5.74.0 - (October 3, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 5.73.2 - (August 10, 2023)
 
 * Changed