+++ packages/terra-application-links/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 6.70.0 - (October 3, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 6.69.1 - (August 4, 2023)
 
 * Changed