+++ packages/terra-popup/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed popup click events propagating to elements underneath it.
+
+## 6.71.0 - (October 3, 2023)
+
 * Added
   * Added aria-label prop for terra-popup to announce the close instruction.